requirements.txt
aiogram==2.25.1
python-dotenv==1.0.1
SQLAlchemy==2.0.31
psycopg2-binary==2.9.9
alembic==1.13.2
uvloop==0.19.0; sys_platform != 'win32'
